+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/dialogs/CommitDialog.java
@@ -59,7 +59,7 @@ import ru.taximaxim.codekeeper.ui.UiSync;
 import ru.taximaxim.codekeeper.ui.differ.DiffTableViewer;
 import ru.taximaxim.codekeeper.ui.localizations.Messages;
 import ru.taximaxim.codekeeper.ui.pgdbproject.PgDbProject;
-import ru.taximaxim.codekeeper.ui.properties.UISettings;
+import ru.taximaxim.codekeeper.ui.settings.UISettings;
 
 public final class CommitDialog extends TrayDialog {